Refined Code

+++ src/com/thoughtworks/vapasi/IFruitRepository.java
@@ -3,7 +3,7 @@ import java.util.HashMap;
 
 public interface IFruitRepository {
        void populatePriceTableFruits();
-       void addFruits(String fruitName, float pricePerKg);
+       void addFruits(String fruitName, Double pricePerKg);
        void removeFruits(String fruitName);
        void showAllFruits();
 }
