+++ src/com/thoughtworks/vapasi/OrderDetail.java
@@ -2,8 +2,8 @@ package com.thoughtworks.vapasi;
 
 public class OrderDetail {
     public int orderDetailId;
-    public float quantity;
-    public float totalPrice;
+    public double quantity;
+    public double totalPrice;
     public String fruit;
     public Order orderId;
